add permission logic for update or delete thread

+++ app/Controllers/Http/ThreadsController.ts
@@ -49,9 +49,17 @@ export default class ThreadsController {
     }
   }
 
-  public async update({ params, request, response }: HttpContextContract) {
+  public async update({ params, auth, request, response }: HttpContextContract) {
     try {
+      const user = await auth.user
       const thread = await Thread.findOrFail(params.id)
+
+      if (thread.userId !== user?.id) {
+        return response
+          .status(403)
+          .json({ message: 'Anda tidak memiliki izin untuk mengedit thread ini.' })
+      }
+
       const validateData = await request.validate(ThreadValidator)
       await thread.merge(validateData).save()
 
@@ -68,16 +76,24 @@ export default class ThreadsController {
     }
   }
 
-  public async destroy({ params, response }: HttpContextContract) {
+  public async destroy({ params, auth, response }: HttpContextContract) {
     try {
+      const user = await auth.user
       const thread = await Thread.findOrFail(params.id)
+
+      if (thread.userId !== user?.id) {
+        return response
+          .status(403)
+          .json({ message: 'Anda tidak memiliki izin untuk mengedit thread ini.' })
+      }
+
       await thread.delete()
       return response.status(200).json({
         message: 'Thread deleted successfully',
       })
     } catch (error) {
       return response.status(400).json({
-        message: error,
+        message: error.message || 'Something went wrong',
       })
     }
   }
